Refactor code and remove no longer needed WeatherIcon component
