--- src/components/Weather/styles.scss
@@ -1,35 +1,0 @@
-@import '../../assets/styles/import.scss';
-@import '../../variables.scss';
-
-.icon {
-    position: relative;
-    display: inline-block;
-    padding-top: 0.5rem;
-    width: 12em;
-    height: 10em;
-    font-size: 0.25em; /* control icon size here */
-
-    &:active {
-        pointer-events: none;
-    }
-}
-
-.icon-entur {
-    position: relative;
-    display: inline-block;
-    font-size: 5.5rem;
-
-    &:active {
-        pointer-events: none;
-    }
-
-    &--sun {
-        animation: spin 30s linear infinite;
-    }
-}
-
-@keyframes spin {
-    100% {
-        transform: rotate(360deg);
-    }
-}
